+++ .build/Publish.Proget.build.ps1
@@ -89,17 +89,6 @@ param
 )
 
 Task publish_module_to_proget -if ($PSTOOLS_APITOKEN) {
-    # This task fails when publishing a non-preview version. The reason is that Publish-PSResource
-    # does not allow the manifest attribute PreRelease = ''. It will fail with a strange error
-    # The required element version is missing in manifest.
-    # The module ModuleBuilder that is responsible for generating the module output requires that
-    # the attribute PreRelease is defined as a empty string in the source module manifest regardless
-    # of weather it will be populated or not. PowershellGet v2 does not care if PreRelease is defined
-    # as an empty string when publishing a non-prerelease version however PowershellGet v3 does not
-    # allow it. And PowershellGet v3 is required when publishing to proget.
-    # Proposed solution is to remove the empty prerelease attribute in the manifest before publishing
-    # to proget.
-
     . Set-SamplerTaskVariable
 
     # Remove empty Prerelease property, see note above
@@ -107,16 +96,13 @@ Task publish_module_to_proget -if ($PSTOOLS_APITOKEN) {
     $UpdatedManifest | Set-Content $BuiltModuleManifest
     Write-Build DarkGray 'Removed empty Prerelease property if present'
 
-    Import-Module -Name 'ModuleBuilder' -ErrorAction Stop
+    Import-Module -name 'ModuleBuilder' -ErrorAction Stop
     Write-Build DarkGray 'Imported module ModuleBuilder'
 
     Write-Build DarkGray "`nAbout to publish '$BuiltModuleBase'."
 
-    Import-Module PowershellGet -RequiredVersion 3.0.17 -Force
-    Write-Build DarkGray 'Imported PowershellGet v3'
-
     $RepoGuid = (New-Guid).Guid
-    Register-PSResourceRepository -Name $RepoGuid -Uri $PSTOOLS_SOURCE -Trusted
+    Register-PSResourceRepository -name $RepoGuid -Uri $PSTOOLS_SOURCE -Trusted
     Write-Build DarkGray 'Registered ResourceRepository'
 
     try
@@ -142,6 +128,6 @@ Task publish_module_to_proget -if ($PSTOOLS_APITOKEN) {
     }
     finally
     {
-        Unregister-PSResourceRepository -Name $RepoGuid -Confirm:$false
+        Unregister-PSResourceRepository -name $RepoGuid -Confirm:$false
     }
 }